requirements.txt
mcp>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0